+++ OCEAN/cppdefs.h
@@ -96,7 +96,6 @@
 # define OBC_SOUTH
                       /* Applications */
 # undef  BIOLOGY
-# undef  FLOATS
 # undef  STATIONS
 # undef  PASSIVE_TRACER
 # undef  SEDIMENT
@@ -360,7 +359,7 @@
 /*
 !           Applications:
 !---------------------------------
-! Biology, floats, Stations,
+! Biology, Stations,
 ! Passive tracer, Sediments, BBL
 !---------------------------------
 !
@@ -396,17 +395,6 @@
 #  define DIAGNOSTICS_BIO
 #  if defined DIAGNOSTICS_BIO && defined PISCES
 #   define key_trc_diaadd
-#  endif
-# endif
-                      /*   Lagrangian floats model    */
-# ifdef FLOATS
-#  undef  FLOATS_GLOBAL_ATTRIBUTES
-#  undef  IBM
-#  undef  RANDOM_WALK
-#  ifdef RANDOM_WALK
-#   define DIEL_MIGRATION
-#   define RANDOM_VERTICAL
-#   define RANDOM_HORIZONTAL
 #  endif
 # endif
                       /*   Stations recording    */
@@ -454,7 +442,6 @@
 # define OBC_SOUTH
                       /* Applications */
 # undef  BIOLOGY
-# undef  FLOATS
 # undef  STATIONS
 # undef  PASSIVE_TRACER
 # undef  SEDIMENT
@@ -600,7 +587,7 @@
 /*
 !           Applications:
 !---------------------------------
-! Biology, floats, Stations,
+! Biology, Stations,
 ! Passive tracer, Sediments, BBL
 !---------------------------------
 !
@@ -909,15 +896,6 @@
 # undef  PSOURCE_MASS
 # define ANA_PSOURCE
 # define NS_PERIODIC
-# undef  FLOATS
-# ifdef FLOATS
-#  define RANDOM_WALK
-#  ifdef RANDOM_WALK
-#   define DIEL_MIGRATION
-#   define RANDOM_VERTICAL
-#   define RANDOM_HORIZONTAL
-#  endif
-# endif
 # define NO_FRCFILE
 
 #elif defined SEAMOUNT
@@ -2038,7 +2016,6 @@
 # undef  PSOURCE_MASS
 # define ANA_PSOURCE
 # define NS_PERIODIC
-# undef  FLOATS
 # define NO_FRCFILE
 # define USE_CALENDAR
 # define NEW_S_COORD